% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa.R
\docType{data}
\name{hsa}
\alias{hsa}
\title{Human genome annotations for differentially expressed genes}
\format{An object of class \code{"data.frame"}}
\usage{
data(hsa)
}
\description{
Data from Biomart annotations on human genes
}
\examples{
data(hsa)

}
\keyword{datasets}
