% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmu.R
\docType{data}
\name{mmu}
\alias{mmu}
\title{Mouse genome annotations for differentially expressed genes}
\format{An object of class \code{"data.frame"}}
\usage{
data(mmu)
}
\description{
Data from Biomart annotations on mouse genes
}
\examples{
data(mmu)

}
\keyword{datasets}
