% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_conversion.R
\name{symbol_to_ensembl}
\alias{symbol_to_ensembl}
\title{Function for converting between Ensembl IDs and gene symbols (canonical names)}
\usage{
symbol_to_ensembl(IDs, mart = NULL, human = TRUE)
}
\arguments{
\item{IDs}{either a data frame of counts in which the first column is a vector of gene symbols or ensembl IDs or, simply, a vector}

\item{mart}{a biomart data frame returned from hciR::read_biomart, if available. If set to NULL, the function will read its own biomart}

\item{human}{a logical value specifying if the IDs are human. Defaults to TRUE. FALSE specifies mouse}
}
\value{
Returns either the original counts table with converted gene IDs, or a data frame of one column with the converted IDs
}
\description{
This function will output the converted names of each gene as a data frame variable, either replacing the names in a given data frame or simply returning a column.
}
\examples{
\dontrun{
ensembl_to_symbol(counts, mart = mmu, human = FALSE)
}
}
\author{
Chris Stehn
}
